load new datasets by default
