+++ apps/anomalist/cli.py
@@ -5,9 +5,10 @@ import click
 import structlog
 from joblib import Memory
 from rich_click.rich_command import RichCommand
+from sqlalchemy.engine import Engine
 
 from apps.anomalist.anomalist_api import ANOMALY_TYPE, anomaly_detection
-from etl.db import get_engine, read_sql
+from etl.db import get_engine, production_or_master_engine, read_sql
 from etl.paths import CACHE_DIR
 
 log = structlog.get_logger()
@@ -91,6 +92,12 @@ def cli(
     ```
     $ etl anomalist --anomaly-type gp --dataset-ids 6589
     ```
+
+    **Example 4:** Create anomalies for new datasets
+
+    ```
+    $ etl anomalist --anomaly-type gp
+    ```
     """
     # Convert variable mapping from JSON to dictionary.
     if variable_mapping:
@@ -104,8 +111,15 @@ def cli(
     else:
         variable_mapping_dict = {}
 
-    # Load all variables from given datasets
-    if dataset_ids:
+    # If no variable IDs are given, load all variables from the given datasets.
+    if not variable_ids:
+        assert not dataset_ids, "Cannot specify both dataset IDs and variable IDs."
+
+        # Use new datasets
+        if not dataset_ids:
+            dataset_ids = load_datasets_new_ids(get_engine())
+
+        # Load all variables from given datasets
         assert not variable_ids, "Cannot specify both dataset IDs and variable IDs."
         q = """
         select id from variables
@@ -120,6 +134,27 @@ def cli(
         dry_run=dry_run,
         force=force,
         reset_db=reset_db,
+    )
+
+
+def load_datasets_new_ids(source_engine: Engine) -> list[int]:
+    # Compare against production or staging-site-master
+    target_engine = production_or_master_engine()
+
+    # Get new datasets
+    # TODO: replace by real catalogPath when we have it in MySQL
+    q = """SELECT
+        id,
+        CONCAT(namespace, "/", version, "/", shortName) as catalogPath
+    FROM datasets
+    """
+    source_datasets = read_sql(q, source_engine)
+    target_datasets = read_sql(q, target_engine)
+
+    return list(
+        source_datasets[
+            source_datasets.catalogPath.isin(set(source_datasets["catalogPath"]) - set(target_datasets["catalogPath"]))
+        ]["id"]
     )
