+++ apps/owidbot/anomalist.py
@@ -3,12 +3,11 @@ import time
 from structlog import get_logger
 
 from apps.anomalist.anomalist_api import anomaly_detection
+from apps.anomalist.cli import load_datasets_new_ids
 from apps.wizard.app_pages.anomalist.utils import load_variable_mapping
 from etl import grapher_model as gm
 from etl.config import OWIDEnv
-from etl.db import Engine, read_sql
-
-from .chart_diff import production_or_master_engine
+from etl.db import read_sql
 
 log = get_logger()
 
@@ -17,13 +16,12 @@ def run(branch: str) -> None:
     """Compute all anomalist for new and updated datasets."""
     # Get engines for branch and production
     source_engine = OWIDEnv.from_staging(branch).get_engine()
-    target_engine = production_or_master_engine()
 
     # Create table with anomalist if it doesn't exist
     gm.Anomaly.create_table(source_engine, if_exists="skip")
 
     # Load new dataset ids
-    datasets_new_ids = _load_datasets_new_ids(source_engine, target_engine)
+    datasets_new_ids = load_datasets_new_ids(source_engine)
 
     if not datasets_new_ids:
         log.info("No new datasets found.")
@@ -48,21 +46,3 @@ def run(branch: str) -> None:
     )
 
     log.info("owidbot.anomalist.end", n_variables=len(variable_ids), t=time.time() - t)
-
-
-def _load_datasets_new_ids(source_engine: Engine, target_engine: Engine) -> list[int]:
-    # Get new datasets
-    # TODO: replace by real catalogPath when we have it in MySQL
-    q = """SELECT
-        id,
-        CONCAT(namespace, "/", version, "/", shortName) as catalogPath
-    FROM datasets
-    """
-    source_datasets = read_sql(q, source_engine)
-    target_datasets = read_sql(q, target_engine)
-
-    return list(
-        source_datasets[
-            source_datasets.catalogPath.isin(set(source_datasets["catalogPath"]) - set(target_datasets["catalogPath"]))
-        ]["id"]
-    )
