+++ apps/owidbot/chart_diff.py
@@ -2,8 +2,8 @@ import pandas as pd
 from structlog import get_logger
 
 from apps.wizard.app_pages.chart_diff.chart_diff import ChartDiffsLoader
-from etl.config import ENV_FILE_PROD, OWID_ENV, OWIDEnv, get_container_name
-from etl.db import Engine
+from etl.config import OWIDEnv, get_container_name
+from etl.db import production_or_master_engine
 
 from . import github_utils as gh_utils
 
@@ -65,17 +65,6 @@ def run(branch: str, charts_df: pd.DataFrame) -> str:
     """.strip()
 
     return body
-
-
-def production_or_master_engine() -> Engine:
-    """Return the production engine if available, otherwise connect to staging-site-master."""
-    if OWID_ENV.env_remote == "production":
-        return OWID_ENV.get_engine()
-    elif ENV_FILE_PROD:
-        return OWIDEnv.from_env_file(ENV_FILE_PROD).get_engine()
-    else:
-        log.warning("ENV file doesn't connect to production DB, comparing against staging-site-master")
-        return OWIDEnv.from_staging("master").get_engine()
 
 
 def call_chart_diff(branch: str) -> pd.DataFrame:
